+++ internal/examples/thrift-hello/hello/echo/helloclient/client.go
@@ -25,12 +25,12 @@ package helloclient
 
 import (
 	"context"
-	"reflect"
 	"go.uber.org/thriftrw/wire"
-	"go.uber.org/yarpc/api/transport"
-	"go.uber.org/yarpc/internal/examples/thrift-hello/hello/echo"
-	"go.uber.org/yarpc/encoding/thrift"
 	"go.uber.org/yarpc"
+	"go.uber.org/yarpc/api/transport"
+	"go.uber.org/yarpc/encoding/thrift"
+	"go.uber.org/yarpc/internal/examples/thrift-hello/hello/echo"
+	"reflect"
 )
 
 // Interface is a client for the Hello service.