+++ internal/examples/thrift-keyvalue/keyvalue/kv/keyvalueclient/client.go
@@ -25,12 +25,12 @@ package keyvalueclient
 
 import (
 	"context"
-	"reflect"
 	"go.uber.org/thriftrw/wire"
-	"go.uber.org/yarpc/api/transport"
-	"go.uber.org/yarpc/internal/examples/thrift-keyvalue/keyvalue/kv"
-	"go.uber.org/yarpc/encoding/thrift"
 	"go.uber.org/yarpc"
+	"go.uber.org/yarpc/api/transport"
+	"go.uber.org/yarpc/encoding/thrift"
+	"go.uber.org/yarpc/internal/examples/thrift-keyvalue/keyvalue/kv"
+	"reflect"
 )
 
 // Interface is a client for the KeyValue service.